fix: generate logfile name at start
In the previous version, the name of the logfile was generated
when the logfile was created.
This meant that the timestamp referred to the end of the transfer.
I think it is more intuitive for users if the timestamp in the filename
refers to the start of the transfer.
Hence the reason of this commit