verify size before checksum
Checking the size of files and data objects is faster
than calculating and comparing their checksums.
And if a file has a different size compared to the
data object, the checksum cannot be the same.
I changed the code so the size is always verified first,
and a checksum is only made if the size is correct.
If the size is occasionally wrong (so data needs
to be overwritten), we would expect a performance increase.
Based on suggestion in
https://github.com/jefscheepers/toolbox/issues/9#issuecomment-2273671664.