improve performance of size check before checksum
We were checking the size of the files before checking the checksum,
because the checksum will not match if the size is different.
However, in the pervious version the code the get function was called
twice on the object, once for the size and once for the checksum.
This commit integrates the size check in the checksum comparison
function, so the data object info needs to be fetched just once.